+++ board/questions.py
@@ -1,53 +1,31 @@
 import logging
+from board.pull_questions_in import TileLoader, Tile
 
 NUM_ROUNDS = 2
-
-
-class Tile:
-    def __init__(self, question, answers, r_answer, points):
-        self.question = question  # String
-        self.answers = answers  # list of answers
-        self.r_answer = r_answer  # index of answer in the list of answers
-        self.points = points  # int
-
-    # interface
-    def check_answer(self, player_ans):
-        """
-            Checks player's answer against the actual answer
-
-            Args:
-            playerAns: index of the player's answer in the list of answers
-
-            Returns:
-            user_correct: a boolean representing whether or not the user's selection was correct
-            points: int, number of points the question is worth
-        """
-        user_correct = (player_ans == self.r_answer)
-
-        return user_correct, self.points
 
 
 class Questions:
 
     # default constructor
-    def __init__(self):
-        politics_tiles = {100: Tile('politics question 1', ['a', 'b', 'c'], 2, 100),
-                          200: Tile('politics question 3', ['a', 'b', 'c'], 2, 200),
-                          400: Tile('politics question 3', ['a', 'b', 'c'], 2, 400),
-                          800: Tile('politics question 4', ['a', 'b', 'c'], 2, 800)
-                          }
-
-        food_tiles = {100: Tile('food question 1', ['a', 'b', 'c'], 2, 100),
-                      200: Tile('food question 3', ['a', 'b', 'c'], 2, 200),
-                      400: Tile('food question 3', ['a', 'b', 'c'], 2, 400),
-                      800: Tile('food question 4', ['a', 'b', 'c'], 2, 800)
-                      }
-
-        self.rounds = {1: {"politics": politics_tiles, "food": food_tiles},
-                       2: {"politics2": politics_tiles, "food2": food_tiles}}
-        self.politics_iterator = 1
-
-        self.food_iterator = 1
+    def __init__(self, filename):
+        # # politics_tiles = {100: Tile('politics question 1', ['a', 'b', 'c'], 2, 100),
+        #                   200: Tile('politics question 3', ['a', 'b', 'c'], 2, 200),
+        #                   400: Tile('politics question 3', ['a', 'b', 'c'], 2, 400),
+        #                   800: Tile('politics question 4', ['a', 'b', 'c'], 2, 800)
+        #                   }
+        #
+        # food_tiles = {100: Tile('food question 1', ['a', 'b', 'c'], 2, 100),
+        #               200: Tile('food question 3', ['a', 'b', 'c'], 2, 200),
+        #               400: Tile('food question 3', ['a', 'b', 'c'], 2, 400),
+        #               800: Tile('food question 4', ['a', 'b', 'c'], 2, 800)
+        #               }
+        questions_in = TileLoader(filename)
+        self.rounds = questions_in.rounds
+        # self.rounds = {1: {"politics": politics_tiles, "food": food_tiles},
+        #                2: {"politics2": politics_tiles, "food2": food_tiles}}
+        # self.politics_iterator = 1
+        #
+        # self.food_iterator = 1
         self.no_category = "There are no more questions left in this category. Spin Again"
 
     def is_category_open(self, category, round):
@@ -83,12 +61,18 @@ class Questions:
                     return next_tile
             return False
 
-# a = Questions()
-# a.check_question("politics")
-# a.get_tile("politics")
-# a.get_tile("politics")
-# a.get_tile("politics")
-# a.get_tile("politics")
-# a.get_tile("politics")
 
-# a.get_tile("food")
+# a = Questions()
+# a.is_category_open("politics", 1)
+# rnd1_cats = a.get_open_categories(1)
+# rnd2_cats = a.get_open_categories(2)
+# cat1 = rnd1_cats[0]
+# cat2 = rnd2_cats[-1]
+# a.is_category_open(rnd1_cats, 1)
+#
+# print(rnd1_cats, cat1)
+# print(rnd2_cats, cat2)
+#
+# for i in range(0, len(a.rounds[1][cat1].items()) + 1):
+#     tile = a.get_tile(cat1, 1)
+#     print(tile)
